de-duplicate ProxyMaker
Note that the git history of one of the ProxyMaker files won't be reflected in the surviving class

* rename JvmSubclassInstrumentation to avoid name class with SubclassInstrumentation
* update JvmSubclassInstrumentation to implement SubclassInstrumentation
* remove unused method setProxyHandler() from SubclassInstrumentation
