+++ modules/mockk-agent-android/src/main/kotlin/io/mockk/proxy/android/transformation/AndroidSubclassInstrumentation.kt
@@ -74,11 +74,4 @@ internal class AndroidSubclassInstrumentation(
 
         return abstractMethods.map { it.originalMethod }.toTypedArray()
     }
-
-    override fun setProxyHandler(proxy: Any, handler: MockKInvocationHandler) {
-        if (ProxyBuilder.isProxyClass(proxy::class.java)) {
-            ProxyBuilder.setInvocationHandler(proxy, ProxyInvocationHandler(handler))
-        }
-    }
-
 }
