+++ modules/mockk-agent-api/src/commonMain/kotlin/io/mockk/proxy/common/transformation/TransformationRequest.kt
@@ -10,6 +10,4 @@ data class TransformationRequest(
         type,
         !untransform
     )
-
-
 }
