+++ modules/mockk-agent-api/src/jvmMain/kotlin/io/mockk/proxy/common/ProxyMaker.kt
@@ -13,19 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package io.mockk.proxy.common
 
 import io.mockk.proxy.*
 import io.mockk.proxy.common.transformation.InlineInstrumentation
-import io.mockk.proxy.common.transformation.SubclassInstrumentation
 import io.mockk.proxy.common.transformation.TransformationRequest
+import io.mockk.proxy.common.transformation.SubclassInstrumentation
 import io.mockk.proxy.common.transformation.TransformationType
-import java.lang.ref.SoftReference
 import java.lang.ref.WeakReference
 import java.lang.reflect.Method
 import java.lang.reflect.Modifier
-import java.lang.reflect.Proxy
 
 class ProxyMaker(
     private val log: MockKAgentLogger,
@@ -34,6 +31,7 @@ class ProxyMaker(
     private val instantiator: MockKInstantiatior,
     private val handlers: MutableMap<Any, MockKInvocationHandler>
 ) : MockKProxyMaker {
+
     override fun <T : Any> proxy(
         clazz: Class<T>,
         interfaces: Array<Class<*>>,
@@ -44,31 +42,24 @@ class ProxyMaker(
 
         throwIfNotPossibleToProxy(clazz, interfaces)
 
-        if (clazz.isInterface) {
-            val proxyInstance = Proxy.newProxyInstance(
-                clazz.classLoader,
-                interfaces + clazz,
-                ProxyInvocationHandler(handler)
-            )
+        // Sometimes (e.g. in case of sealed classes) we will create the proxy for a subclass of `clazz` and not `clazz`
+        // itself.  We need to determine this early, so that the subclass will be inlined as well.
+        val actualClass = findActualClassToBeProxied(clazz)
 
-            return CancelableResult(clazz.cast(proxyInstance))
-        }
-
-        val cancellation = inline(clazz)
+        val cancellation = inline(actualClass)
 
         val result = CancelableResult<T>(cancelBlock = cancellation)
 
         val proxyClass = try {
-            subclass(clazz, interfaces)
+            subclass(actualClass, interfaces)
         } catch (ex: Exception) {
             result.cancel()
-            throw MockKAgentException("Failed to subclass $clazz", ex)
+            throw MockKAgentException("Failed to subclass $actualClass", ex)
         }
 
         try {
-            val proxy = instantiate(clazz, proxyClass, useDefaultConstructor, instance)
+            val proxy = instantiate(actualClass, proxyClass, useDefaultConstructor, instance)
 
-            subclasser.setProxyHandler(proxy, handler)
             handlers[proxy] = handler
             val callbackRef = WeakReference(proxy)
             return result
@@ -77,11 +68,9 @@ class ProxyMaker(
                     callbackRef.get()?.let {
                         handlers.remove(it)
                     }
-
                 }
         } catch (e: Exception) {
             result.cancel()
-
             throw MockKAgentException("Instantiation exception", e)
         }
     }
@@ -114,11 +103,7 @@ class ProxyMaker(
         val superclasses = getAllSuperclasses(clazz)
 
         return if (inliner != null) {
-            val transformRequest =
-                TransformationRequest(
-                    superclasses,
-                    TransformationType.SIMPLE
-                )
+            val transformRequest = TransformationRequest(superclasses, TransformationType.SIMPLE)
 
             inliner.execute(transformRequest)
         } else {
@@ -130,12 +115,30 @@ class ProxyMaker(
         }
     }
 
+    private fun <T : Any> findActualClassToBeProxied(
+        clazz: Class<T>,
+    ): Class<T> {
+        val kClass = clazz.kotlin
+        if (!kClass.isSealed) {
+            return clazz
+        }
+
+        val subclass = kClass.sealedSubclasses.firstOrNull()?.java
+            ?: error("Unable to create proxy for sealed class $clazz, no subclasses available")
+        log.trace("Class $clazz is sealed, will use its subclass $subclass to build proxy")
+        @Suppress("UNCHECKED_CAST")
+        return findActualClassToBeProxied(subclass) as Class<T>
+    }
+
     private fun <T : Any> subclass(
         clazz: Class<T>,
         interfaces: Array<Class<*>>
     ): Class<T> {
         return if (Modifier.isFinal(clazz.modifiers)) {
             log.trace("Taking instance of $clazz itself because it is final.")
+            clazz
+        } else if (interfaces.isEmpty() && !Modifier.isAbstract(clazz.modifiers) && inliner != null) {
+            log.trace("Taking instance of $clazz itself because it is not abstract and no additional interfaces specified.")
             clazz
         } else {
             log.trace(
@@ -175,7 +178,7 @@ class ProxyMaker(
             val defaultConstructor = cls.getDeclaredConstructor()
             try {
                 defaultConstructor.isAccessible = true
-            } catch (ex: Exception) {
+            } catch (ignored: Exception) {
                 // skip
             }
 
@@ -200,6 +203,7 @@ class ProxyMaker(
 
 
     companion object {
+
         private val notMockableClasses = setOf(
             Class::class.java,
             Boolean::class.java,
